+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.18
 
+- Show times in the events timezone
 - Fix to prevent repeated messaging about geolocation when user denies access
 - Put time into the event title rather than after the description